visomaster_modificado/requirements_cu118.txt
--extra-index-url https://download.pytorch.org/whl/cu118

numpy>=1.26.0
opencv-python>=4.8.1.78
scikit-image==0.21.0
pillow==9.5.0
onnx==1.16.1
protobuf==4.23.2
psutil>=5.9.7
onnxruntime-gpu==1.18.0
packaging==24.1
PySide6>=6.6.1
kornia
torch>=2.1.1
torchvision>=0.16.1
torchaudio==2.4.1+cu118
tqdm
scipy
insightface==0.7.3
huggingface-hub==0.23.2
gradio
onnxsim
requests
pyqttoast>=1.0.0
qdarkstyle
pyqtdarktheme
insightface>=0.7.3
onnxruntime>=1.16.0
gfpgan>=1.3.8
typing_extensions>=4.8.0
deoldify-th>=1.0.0
matplotlib>=3.7.2